refactor(LottoController) : 패키지 이동
